Add Ansible Module support to schema

+++ src/main/java/ir/msdehghan/plugins/ansible/model/ansible/play/LoopControl.java
@@ -19,7 +19,7 @@ public class LoopControl extends YamlMappingType {
                 .setType(STRING);
 
         addField("pause")
-                .setType(INTEGER) // TODO: float
+                .setType(FLOAT)
                 .setDescription("Default: 0");
 
         addField("extended")
